delete-me: Disable mimalloc and force debug symbol for release builds
We remove mimalloc to use the system malloc and let Instruments measure
the memory usage of Meilisearch. The debug symbols are forced in release
for Instruments to have much more information on the program.

Please, don't use these parameters in a Meilisearch release as they are
making it very very very slow.